requirements.txt
streamlit
groq
tavily-python
langgraph
pydantic
beautifulsoup4
requests
pypdf
python-docx